Refactor and add pinataIpfs into the config.
