+++ packages/tznft/src/config/parser.ts
@@ -11,12 +11,18 @@ const network = z.object({
   aliases: z.record(alias)
 });
 
-export const configStorage = z.object({
+const pinataIpfs = z.object({
+  apiKey: z.string(),
+  secretKey: z.string()
+});
+
+export const config = z.object({
   activeNetwork: z.string(),
-  availableNetworks: z.record(network)
+  availableNetworks: z.record(network),
+  pinataIpfs: pinataIpfs.optional()
 });
 
 export type Alias = z.infer<typeof alias>;
 export type Network = z.infer<typeof network>;
-export type ConfigStorage = z.infer<typeof configStorage>;
-export type Config = Network;
+export type PinataIpfs = z.infer<typeof pinataIpfs>;
+export type Config = z.infer<typeof config>;
